multispecies implementation

See merge request atomicclusterexpansion/lammps-ace!28

+++ src/USER-PACE/ace_types.h
@@ -39,7 +39,8 @@ typedef short int DENSITY_TYPE;
 
 typedef short int MS_TYPE;
 
-typedef short int SHORT_INT_TYPE;
+typedef int SHORT_INT_TYPE;
 typedef double DOUBLE_TYPE;
 
 #endif
+
